agent_results/end2end_gemini/iclr2025_bi_align/claude_code/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
torch==2.0.1
transformers==4.30.2
openai==1.2.0
anthropic==0.5.0
tqdm==4.66.1
PyYAML==6.0.1
python-dotenv==1.0.0
scipy==1.11.1
statsmodels==0.14.0